ER-426 :tractor: - Moving vacancy view controller and orchestrator into VacancyViewController and VacancyViewOrchestrator.
