+++ src/Employer/Employer.Web/ViewModels/VacancyView/ApprovedVacancyViewModel.cs
