+++ src/Employer/Employer.Web/ViewModels/VacancyView/ClosedVacancyViewModel.cs
