+++ src/Employer/Employer.Web/Views/VacancyView/ClosedWithApplications.cshtml
