+++ src/Employer/Employer.Web/Views/VacancyView/FullVacancyView.cshtml
