+++ src/Employer/Employer.Web/Views/VacancyView/Submitted.cshtml
